+++ packages/native_animated_image/lib/native_animated_image.dart
@@ -14,3 +14,5 @@ export 'src/ffi/native_animated_image_ffi.dart'
         AnimatedFrame,
         NativeAnimatedImageException,
         NativeAnimatedImageFfi;
+export 'src/platform/native_avif_platform.dart'
+    show NativeAvifPlatform, NativeAvifPlatformException;